+++ torchtext/datasets/amazonreviewfull.py
@@ -1,8 +1,8 @@
 from torchtext.utils import unicode_csv_reader
-from torchtext.data._datasets_utils import _RawTextIterableDataset
-from torchtext.data._datasets_utils import _wrap_split_argument
-from torchtext.data._datasets_utils import _add_docstring_header
-from torchtext.data._datasets_utils import _download_extract_validate
+from torchtext.data.datasets_utils import _RawTextIterableDataset
+from torchtext.data.datasets_utils import _wrap_split_argument
+from torchtext.data.datasets_utils import _add_docstring_header
+from torchtext.data.datasets_utils import _download_extract_validate
 import io
 import os
 import logging
